SUMMARY:   Redefined all HCURL basis functions to use edge tangent scaled by edge length in the DoF definition. Changed tests. Redefined HDIV basis on HEX to use face outer normal scaled by face area. Changed corresponding test.
     BUGS:      no

     CHECKLIST:

     1)   Did you add any source files?  No
     1.a) Did you do a 'cvs add'?  [???]
     1.b) Did you add them to a Makefile.am file in some way?  [???]
     1.c) Did you do a new bootstrap?  Yes

     2)   Did you update or change the build system?  [???]
     2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
          Makefile.am files and the Makefile.in files?  [???]
     2.b) [Optional but recommended] Did you blow away the libraries and
          executables and then rebuild?  [???]

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes

     COMMENTS:



(checklist version 2007-06-20)

+++ packages/intrepid/src/Cell/Intrepid_CellToolsDef.hpp
@@ -1357,7 +1357,7 @@ void CellTools<Scalar>::getPhysicalSideNormals(ArrayTypeOut &                sid
                       ">>> ERROR (Intrepid::CellTools::getPhysicalSideNormals): two or three-dimensional parent cell required");
   
   // Check side ordinal: by definition side is subcell whose dimension = spaceDim-1
-  TEST_FOR_EXCEPTION( !( (0 <= worksetSideOrd) && (worksetSideOrd < parentCell.getSubcellCount(spaceDim - 1) ) ), std::invalid_argument,
+  TEST_FOR_EXCEPTION( !( (0 <= worksetSideOrd) && (worksetSideOrd < (int)parentCell.getSubcellCount(spaceDim - 1) ) ), std::invalid_argument,
                       ">>> ERROR (Intrepid::CellTools::getPhysicalSideNormals): side ordinal out of bounds");  
 #endif
